Split `ListenBrainz.cs` into multiple parts
`ListenBrainz.Internals.cs` contains all the internals,
`ListenBrainz.EndPoints.cs` contains the public API for the various API
endpoints, and `ListenBrainz.cs` contains the constructors, properties
and `HttpClient`/`IDispose` handling.
